images/streamlit/requirements.txt
boto3
langchain-aws
langchain-chroma
langchain-community
sentence-transformers
langchain-ollama
pytest
pyyaml
gitpython
pypdf
streamlit
chromadb
langchain-huggingface
